# contracts/openapi/central.openapi.json
{
...
  "info": {
    "description": "Central API for global administration, cross-store reporting, integrations, and Store Edge synchronization.",
    "title": "Mercadia Central Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/stores/{storeId}/sync-events": {
      "post": {
        "operationId": "acceptStoreSyncEvents",
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Unique key used to make command execution idempotent.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "accepted": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string"
                    },
                    "storeId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "storeId",
                    "status",
                    "accepted"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Sync batch accepted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Invalid sync batch"
          }
        },
        "summary": "Accept synchronized Store Edge events",
        "tags": [
          "sync"
        ]
      }
    }
  }
}